services/api_gateway_service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
slowapi==0.1.9
pydantic==2.11.7
